# docs/swagger.yaml
definitions:
  api.Response:
    properties:
      code:
        type: integer
      data: {}
      message:
        type: string
    type: object
  stream.CreateStreamRequest:
    properties:
      description:
        example: This is my new stream, check it out
        type: string
      name:
        example: My New Stream
        type: string
    required:
    - name
    type: object
...
  stream.Stream:
    properties:
      created_by:
        type: integer
      created_date:
        type: string
      description:
        type: string
      endDate:
        type: string
      id:
        type: integer
      manifest_path:
        type: string
      name:
        type: string
      slug:
        type: string
      startDate:
        type: string
      updated_by:
        type: integer
      updated_date:
        type: string
    required:
    - name
    - slug
    type: object
  streams.response:
    properties:
      created_by:
        type: integer
      created_date:
        type: string
      description:
        type: string
      endDate:
        type: string
      id:
        type: integer
      is_live:
        type: boolean
      manifest_path:
        type: string
      name:
        type: string
      slug:
        type: string
      startDate:
        type: string
      updated_by:
        type: integer
      updated_date:
        type: string
    required:
    - name
    - slug
    type: object
info:
  contact: {}
paths:
...
  /v1/streams/{id}:
    delete:
      consumes:
      - application/json
      description: Delete a stream from the database
      parameters:
      - description: Stream ID to delete
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/api.Response'
            - properties:
                data:
                  $ref: '#/definitions/stream.Stream'
              type: object
        "400":
          description: Bad Request
          schema:
            allOf:
            - $ref: '#/definitions/api.Response'
            - properties:
                data:
                  type: string
              type: object
      summary: Delete Stream
      tags:
      - stream
    get:
      consumes:
      - application/json
      description: Get a stream by their ID
      parameters:
      - description: Stream ID to get
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/api.Response'
            - properties:
                data:
                  $ref: '#/definitions/streams.response'
              type: object
        "400":
          description: Bad Request
          schema:
            allOf:
            - $ref: '#/definitions/api.Response'
            - properties:
                data:
                  type: string
              type: object
      summary: Get Stream
      tags:
      - stream
    put:
      consumes:
      - application/json
      description: Update a stream by their ID
      parameters:
      - description: Stream ID to get
        in: path
        name: id
        required: true
        type: integer
      - description: Body Request
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/stream.CreateStreamRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/api.Response'
            - properties:
                data:
                  $ref: '#/definitions/stream.Stream'
              type: object
        "400":
          description: Bad Request
          schema:
            allOf:
            - $ref: '#/definitions/api.Response'
            - properties:
                data:
                  type: string
              type: object
      summary: Update Stream
      tags:
      - stream
...
swagger: "2.0"